Extracted pan and zoom in a shared component
