+++ src/h5web/visualizations/heatmap/hooks.ts
@@ -1,14 +1,4 @@
-import React, {
-  useCallback,
-  useRef,
-  useEffect,
-  CSSProperties,
-  useContext,
-  useMemo,
-} from 'react';
-import { Vector3 } from 'three';
-import { ReactThreeFiber, PointerEvent, useThree } from 'react-three-fiber';
-import { clamp } from 'lodash-es';
+import { CSSProperties, useContext, useMemo, useEffect } from 'react';
 import { useComlink } from 'react-use-comlink';
 import { useMeasure, useSetState } from 'react-use';
 import shallow from 'zustand/shallow';
@@ -21,8 +11,6 @@ import { TextureWorker } from './worker';
 // eslint-disable-next-line
 // @ts-ignore
 import Worker from 'worker-loader!./worker'; // eslint-disable-line
-
-const ZOOM_FACTOR = 0.95;
 
 export function useProps(): HeatmapProps {
   const props = useContext(HeatmapContext);
@@ -147,99 +135,4 @@ export function useHeatmapStyles<T>(): [
       paddingLeft: leftAxisWidth,
     },
   ];
-}
-
-export function usePanZoom(): ReactThreeFiber.Events {
-  const { camera, invalidate, size } = useThree();
-  const { width, height } = size;
-
-  const startOffsetPosition = useRef<Vector3>(); // `useRef` to avoid re-renders
-
-  const moveCameraTo = useCallback(
-    (x: number, y: number) => {
-      const { position, zoom } = camera;
-
-      const factor = (1 - 1 / zoom) / 2;
-      const xBound = width * factor;
-      const yBound = height * factor;
-
-      position.set(
-        clamp(x, -xBound, xBound),
-        clamp(y, -yBound, yBound),
-        position.z
-      );
-
-      invalidate();
-    },
-    [camera, height, invalidate, width]
-  );
-
-  const onPointerDown = useCallback(
-    (evt: PointerEvent) => {
-      const { currentTarget, pointerId } = evt as React.PointerEvent;
-      currentTarget.setPointerCapture(pointerId);
-
-      const projectedPoint = camera.worldToLocal(evt.unprojectedPoint.clone());
-      startOffsetPosition.current = camera.position.clone().add(projectedPoint);
-    },
-    [camera]
-  );
-
-  const onPointerUp = useCallback((evt: PointerEvent) => {
-    const { currentTarget, pointerId } = evt as React.PointerEvent;
-    currentTarget.releasePointerCapture(pointerId);
-
-    startOffsetPosition.current = undefined;
-  }, []);
-
-  const onPointerMove = useCallback(
-    (evt: PointerEvent) => {
-      if (!startOffsetPosition.current) {
-        return;
-      }
-
-      // Prevent events from reaching tooltip mesh when panning
-      evt.stopPropagation();
-
-      const projectedPoint = camera.worldToLocal(evt.unprojectedPoint.clone());
-      const { x: pointerX, y: pointerY } = projectedPoint;
-      const { x: startX, y: startY } = startOffsetPosition.current;
-
-      moveCameraTo(startX - pointerX, startY - pointerY);
-    },
-    [camera, moveCameraTo]
-  );
-
-  const onWheel = useCallback(
-    (evt: PointerEvent) => {
-      const { deltaY } = evt as React.WheelEvent;
-      const factor = deltaY > 0 ? ZOOM_FACTOR : 1 / ZOOM_FACTOR;
-
-      // eslint-disable-next-line no-param-reassign
-      camera.zoom = Math.max(1, camera.zoom * factor);
-      camera.updateProjectionMatrix();
-
-      const projectedPoint = camera.worldToLocal(evt.unprojectedPoint.clone());
-      const { x: pointerX, y: pointerY } = projectedPoint;
-      const { x: camX, y: camY } = camera.position;
-
-      moveCameraTo(
-        camX + pointerX * (1 - 1 / factor),
-        camY + pointerY * (1 - 1 / factor)
-      );
-    },
-    [camera, moveCameraTo]
-  );
-
-  useEffect(() => {
-    // Move camera on resize to stay within mesh bounds
-    moveCameraTo(camera.position.x, camera.position.y);
-  }, [camera, moveCameraTo]);
-
-  return {
-    onPointerDown,
-    onPointerUp,
-    onPointerMove,
-    onWheel,
-  };
 }
