+++ src/h5web/visualizations/shared/PanZoomMesh.tsx
@@ -1,0 +1,102 @@
+import React, { useRef, useCallback, useEffect } from 'react';
+import { Vector3 } from 'three';
+import { PointerEvent, useThree } from 'react-three-fiber';
+import { clamp } from 'lodash-es';
+
+const ZOOM_FACTOR = 0.95;
+
+function PanZoomMesh(): JSX.Element {
+  const { camera, invalidate, size } = useThree();
+  const { width, height } = size;
+
+  const startOffsetPosition = useRef<Vector3>(); // `useRef` to avoid re-renders
+
+  const moveCameraTo = useCallback(
+    (x: number, y: number) => {
+      const { position, zoom } = camera;
+
+      const factor = (1 - 1 / zoom) / 2;
+      const xBound = width * factor;
+      const yBound = height * factor;
+
+      position.set(
+        clamp(x, -xBound, xBound),
+        clamp(y, -yBound, yBound),
+        position.z
+      );
+
+      invalidate();
+    },
+    [camera, height, invalidate, width]
+  );
+
+  const onPointerDown = useCallback(
+    (evt: PointerEvent) => {
+      const { currentTarget, pointerId } = evt as React.PointerEvent;
+      currentTarget.setPointerCapture(pointerId);
+
+      const projectedPoint = camera.worldToLocal(evt.unprojectedPoint.clone());
+      startOffsetPosition.current = camera.position.clone().add(projectedPoint);
+    },
+    [camera]
+  );
+
+  const onPointerUp = useCallback((evt: PointerEvent) => {
+    const { currentTarget, pointerId } = evt as React.PointerEvent;
+    currentTarget.releasePointerCapture(pointerId);
+
+    startOffsetPosition.current = undefined;
+  }, []);
+
+  const onPointerMove = useCallback(
+    (evt: PointerEvent) => {
+      if (!startOffsetPosition.current) {
+        return;
+      }
+
+      // Prevent events from reaching tooltip mesh when panning
+      evt.stopPropagation();
+
+      const projectedPoint = camera.worldToLocal(evt.unprojectedPoint.clone());
+      const { x: pointerX, y: pointerY } = projectedPoint;
+      const { x: startX, y: startY } = startOffsetPosition.current;
+
+      moveCameraTo(startX - pointerX, startY - pointerY);
+    },
+    [camera, moveCameraTo]
+  );
+
+  const onWheel = useCallback(
+    (evt: PointerEvent) => {
+      const { deltaY } = evt as React.WheelEvent;
+      const factor = deltaY > 0 ? ZOOM_FACTOR : 1 / ZOOM_FACTOR;
+
+      // eslint-disable-next-line no-param-reassign
+      camera.zoom = Math.max(1, camera.zoom * factor);
+      camera.updateProjectionMatrix();
+
+      const projectedPoint = camera.worldToLocal(evt.unprojectedPoint.clone());
+      const { x: pointerX, y: pointerY } = projectedPoint;
+      const { x: camX, y: camY } = camera.position;
+
+      moveCameraTo(
+        camX + pointerX * (1 - 1 / factor),
+        camY + pointerY * (1 - 1 / factor)
+      );
+    },
+    [camera, moveCameraTo]
+  );
+
+  useEffect(() => {
+    // Move camera on resize to stay within mesh bounds
+    moveCameraTo(camera.position.x, camera.position.y);
+  }, [camera, moveCameraTo]);
+
+  return (
+    <mesh {...{ onPointerMove, onPointerUp, onPointerDown, onWheel }}>
+      <planeBufferGeometry attach="geometry" args={[width, height]} />
+    </mesh>
+  );
+}
+
+export default PanZoomMesh;
